requirements.txt
streamlit
rembg
onnxruntime
Pillow